More CSS tweaks, added a lib/. assets folder, moved internal help to htm/
In this context a "htm" file in lib/htm/some.htm is a "HyperText Module"
where the content is wrapped with an element with a unique ID and may
include it's own style and script tags. The idea is that it can be a self
contained "module" that tries not to require the main global page space
to know about included modules in advance.