Added localStorage caching and pushState
Moved the edit related links to a kind of fixed dropdown menu, or panel, so
they wouldn't take up so uch room across the top. The new STATE link is simply
for demonstrating the pushState() functionality and an example of how to switch
states between the EDITOR and STATES links.